group3-requirements.txt
streamlit==1.25.0
streamlit-option-menu==0.3.2
opencv-python==4.5.5.64
numpy==1.24.2
Pillow==9.5.0